+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 31-10-2022
+### Added
+- Unit tests for Block, Model and Controller classes.
+
+### Changed
+- Made minor change in code to fix php8.1 compatibility issue.
+- Updated return type of function.
+
+### Removed
+- Removed leftover docblock params for a function.q
+
 ## [2.0.1] - 30-10-2022
 ### Added
 - Backup preview now show title and identifier.